Updated for new patch
hunt2 is modified as well to a working path - this may not be the best path, but it works.  It also requires Deepstrike which is lower in the rage tree so this might not work for everyone.  Even though this is probably not the best path, all the new unstoppable stuff works.