Update Hunt2 and Added stateReset to package
Not sure when stateReset became a thing, but it's great.  I redefined it in Fury pack(we can do this for all of em if we want).  System should track if mob is bleeding now, and hunt2 path will take advantage of bleeding to do sear and gutrend if target is bleeding.